+++ backend/canisters/notifications_index/impl/src/lib.rs
@@ -17,6 +17,7 @@ use utils::fcm_token_store::FcmTokenStore;
 use utils::idempotency_checker::IdempotencyChecker;
 
 mod guards;
+mod jobs;
 mod lifecycle;
 mod memory;
 mod model;